Remote instance feature
