+++ docs/mwdblib.rst
@@ -5,6 +5,10 @@ mwdblib core interface
 
 .. autoclass:: MWDB
     :members:
+    :inherited-members:
 
 .. autoclass:: APIClient
+    :members:
+
+.. autoclass:: mwdblib.remote.MWDBRemote
     :members:
