+++ src/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "3.4.0"
+__version__ = "3.5.0"
